+++ test/integration/_.js
@@ -6,6 +6,7 @@
 import * as arbitrary from "../_arbitraries.js";
 import * as constants from "../_constants.cjs";
 import * as macros from "../_macros.js";
+import * as common from "./_common.js";
 import * as generate from "./_generators.js";
 
-export { arbitrary, constants, generate, macros };
+export { common, arbitrary, constants, generate, macros };