+++ test/integration/escape-all/commonjs.test.js
@@ -1,0 +1,39 @@
+/**
+ * @overview Contains integration test for the CommonJS version of
+ * `Shescape#escapeAll`.
+ * @license MIT
+ */
+
+import { testProp } from "@fast-check/ava";
+import * as fc from "fast-check";
+
+import { arbitrary } from "../_.js";
+
+import { Shescape } from "shescape";
+import { Shescape as ShescapeCjs } from "../../../index.cjs";
+
+testProp(
+  "esm === cjs",
+  [fc.array(arbitrary.shescapeArg()), arbitrary.shescapeOptions()],
+  (t, args, options) => {
+    let shescapeEsm, resultEsm, errorEsm;
+    let shescapeCjs, resultCjs, errorCjs;
+
+    try {
+      shescapeEsm = new Shescape(options);
+      resultEsm = shescapeEsm.escapeAll(args);
+    } catch (error) {
+      errorEsm = error;
+    }
+
+    try {
+      shescapeCjs = new ShescapeCjs(options);
+      resultCjs = shescapeCjs.escapeAll(args);
+    } catch (error) {
+      errorCjs = error;
+    }
+
+    t.deepEqual(resultEsm, resultCjs);
+    t.deepEqual(errorEsm, errorCjs);
+  },
+);