+++ test/integration/escape-all/powershell-exe.test.js
@@ -1,0 +1,20 @@
+/**
+ * @overview Contains integration tests for `Shescape#escapeAll` for Windows
+ * PowerShell (with extension).
+ * @license MIT
+ */
+
+import { common, constants, generate } from "../_.js";
+
+import { Shescape } from "shescape";
+
+const runTest = common.getTestFn(constants.binPowerShell);
+
+runTest(`input is escaped for ${constants.binPowerShell}`, (t) => {
+  for (const scenario of generate.escapeExamples(constants.binPowerShell)) {
+    const { expected, input, options } = scenario;
+    const shescape = new Shescape(options);
+    const result = shescape.escapeAll([input]);
+    t.deepEqual(result, [expected]);
+  }
+});