+++ test/integration/escape-all/valid.test.js
@@ -1,32 +1,14 @@
 /**
- * @overview Contains integration tests for `Shescape#escapeAll`.
+ * @overview Contains integration tests for valid use of `Shescape#escapeAll`.
  * @license MIT
  */
 
 import { testProp } from "@fast-check/ava";
-import test from "ava";
 import * as fc from "fast-check";
 
-import { arbitrary, constants, generate } from "./_.js";
+import { arbitrary } from "../_.js";
 
 import { Shescape } from "shescape";
-import { Shescape as ShescapeCjs } from "../../index.cjs";
-
-for (const shell of generate.platformShells()) {
-  test(`inputs are escaped for ${shell}`, (t) => {
-    for (const { expected, input, options } of generate.escapeExamples(shell)) {
-      let shescape;
-      try {
-        shescape = new Shescape(options);
-      } catch (_) {
-        return t.pass();
-      }
-
-      const result = shescape.escapeAll([input]);
-      t.deepEqual(result, [expected]);
-    }
-  });
-}
 
 testProp(
   "return values",
@@ -100,44 +82,5 @@ testProp(
     }
 
     t.throws(() => shescape.escapeAll(arg), { instanceOf: TypeError });
-  },
-);
-
-testProp("invalid arguments", [arbitrary.shescapeOptions()], (t, options) => {
-  let shescape;
-  try {
-    shescape = new Shescape(options);
-  } catch (_) {
-    return t.pass();
-  }
-
-  for (const { value } of constants.illegalArguments) {
-    t.throws(() => shescape.escapeAll([value]), { instanceOf: TypeError });
-  }
-});
-
-testProp(
-  "esm === cjs",
-  [fc.array(arbitrary.shescapeArg()), arbitrary.shescapeOptions()],
-  (t, args, options) => {
-    let shescapeEsm, resultEsm, errorEsm;
-    let shescapeCjs, resultCjs, errorCjs;
-
-    try {
-      shescapeEsm = new Shescape(options);
-      resultEsm = shescapeEsm.escapeAll(args);
-    } catch (error) {
-      errorEsm = error;
-    }
-
-    try {
-      shescapeCjs = new ShescapeCjs(options);
-      resultCjs = shescapeCjs.escapeAll(args);
-    } catch (error) {
-      errorCjs = error;
-    }
-
-    t.deepEqual(resultEsm, resultCjs);
-    t.deepEqual(errorEsm, errorCjs);
   },
 );