+++ test/integration/quote-all/valid.test.js
@@ -1,51 +1,14 @@
 /**
- * @overview Contains integration tests for `Shescape#quoteAll`.
+ * @overview Contains integration tests for valid use of `Shescape#quoteAll`.
  * @license MIT
  */
 
 import { testProp } from "@fast-check/ava";
-import test from "ava";
 import * as fc from "fast-check";
 
-import { arbitrary, constants, generate } from "./_.js";
+import { arbitrary } from "../_.js";
 
 import { Shescape } from "shescape";
-import { Shescape as ShescapeCjs } from "../../index.cjs";
-
-for (const shell of generate.platformShells()) {
-  if (shell === false) {
-    continue;
-  }
-
-  test(`inputs are quoted for ${shell}`, (t) => {
-    for (const { expected, input, options } of generate.quoteExamples(shell)) {
-      let shescape;
-      try {
-        shescape = new Shescape(options);
-      } catch (_) {
-        return t.pass();
-      }
-
-      const result = shescape.quoteAll([input]);
-      t.deepEqual(result, [expected]);
-    }
-  });
-}
-
-testProp(
-  "quote without shell",
-  [
-    fc.oneof(
-      arbitrary.shescapeArg(),
-      fc.array(arbitrary.shescapeArg(), { minLength: 1 }),
-    ),
-    arbitrary.shescapeOptions().filter((options) => options?.shell === false),
-  ],
-  (t, args, options) => {
-    const shescape = new Shescape(options);
-    t.throws(() => shescape.quoteAll(args), { instanceOf: Error });
-  },
-);
 
 testProp(
   "quote with shell",
@@ -128,44 +91,5 @@ testProp(
     }
 
     t.throws(() => shescape.quoteAll(arg), { instanceOf: TypeError });
-  },
-);
-
-testProp("invalid arguments", [arbitrary.shescapeOptions()], (t, options) => {
-  let shescape;
-  try {
-    shescape = new Shescape(options);
-  } catch (_) {
-    return t.pass();
-  }
-
-  for (const { value } of constants.illegalArguments) {
-    t.throws(() => shescape.quoteAll([value]), { instanceOf: TypeError });
-  }
-});
-
-testProp(
-  "esm === cjs",
-  [fc.array(arbitrary.shescapeArg()), arbitrary.shescapeOptions()],
-  (t, args, options) => {
-    let shescapeEsm, resultEsm, errorEsm;
-    let shescapeCjs, resultCjs, errorCjs;
-
-    try {
-      shescapeEsm = new Shescape(options);
-      resultEsm = shescapeEsm.quoteAll(args);
-    } catch (error) {
-      errorEsm = error;
-    }
-
-    try {
-      shescapeCjs = new ShescapeCjs(options);
-      resultCjs = shescapeCjs.quoteAll(args);
-    } catch (error) {
-      errorCjs = error;
-    }
-
-    t.deepEqual(resultEsm, resultCjs);
-    t.deepEqual(errorEsm, errorCjs);
   },
 );